LPS-54053 Osgsi modules can't share the same packages
