+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/translator/test/ClassicToCreoleTranslatorTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.translator;
+package com.liferay.wiki.translator.test;
 
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
