+++ modules/apps/wiki/wiki-test/test/integration/com/liferay/wiki/trash/test/WikiPageTrashHandlerTest.java
@@ -12,23 +12,22 @@
  * details.
  */
 
-package com.liferay.wiki.trash;
+package com.liferay.wiki.trash.test;
 
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.model.BaseModel;
 import com.liferay.portal.model.ClassedModel;
+import com.liferay.portal.model.Group;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portlet.trash.test.BaseTrashHandlerTestCase;
-import com.liferay.portlet.trash.util.TrashUtil;
+import com.liferay.wiki.asset.WikiPageAssetRenderer;
 import com.liferay.wiki.model.WikiNode;
-import com.liferay.wiki.service.WikiNodeLocalServiceUtil;
+import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.util.test.WikiPageTrashHandlerTestUtil;
 
 import org.junit.ClassRule;
 import org.junit.Ignore;
@@ -37,9 +36,10 @@ import org.junit.Test;
 
 /**
  * @author Eudaldo Alonso
+ * @author Roberto Díaz
  */
 @Sync
-public class WikiNodeTrashHandlerTest extends BaseTrashHandlerTestCase {
+public class WikiPageTrashHandlerTest extends BaseTrashHandlerTestCase {
 
 	@ClassRule
 	@Rule
@@ -51,12 +51,6 @@ public class WikiNodeTrashHandlerTest extends BaseTrashHandlerTestCase {
 	@Ignore()
 	@Override
 	@Test
-	public void testDeleteTrashVersions() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
 	public void testTrashAndDeleteDraft() throws Exception {
 	}
 
@@ -64,25 +58,6 @@ public class WikiNodeTrashHandlerTest extends BaseTrashHandlerTestCase {
 	@Override
 	@Test
 	public void testTrashAndRestoreDraft() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
-	public void testTrashBaseModelAndParentAndDeleteGroupTrashEntries()
-		throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
-	public void testTrashBaseModelAndParentAndDeleteParent() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
-	public void testTrashBaseModelAndParentAndRestoreModel() throws Exception {
 	}
 
 	@Ignore()
@@ -137,31 +112,7 @@ public class WikiNodeTrashHandlerTest extends BaseTrashHandlerTestCase {
 	@Ignore()
 	@Override
 	@Test
-	public void testTrashParentAndDeleteGroupTrashEntries() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
-	public void testTrashParentAndDeleteParent() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
 	public void testTrashRecentBaseModel() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
-	public void testTrashVersionBaseModelAndDelete() throws Exception {
-	}
-
-	@Ignore()
-	@Override
-	@Test
-	public void testTrashVersionBaseModelAndRestore() throws Exception {
 	}
 
 	@Ignore()
@@ -182,69 +133,91 @@ public class WikiNodeTrashHandlerTest extends BaseTrashHandlerTestCase {
 			ServiceContext serviceContext)
 		throws Exception {
 
-		serviceContext = (ServiceContext)serviceContext.clone();
+		return WikiPageTrashHandlerTestUtil.addBaseModelWithWorkflow(
+			parentBaseModel, approved, serviceContext);
+	}
 
-		serviceContext.setWorkflowAction(WorkflowConstants.ACTION_PUBLISH);
-
-		return WikiNodeLocalServiceUtil.addNode(
-			TestPropsValues.getUserId(), getSearchKeywords(),
-			RandomTestUtil.randomString(), serviceContext);
+	@Override
+	protected Long getAssetClassPK(ClassedModel classedModel) {
+		return WikiPageAssetRenderer.getClassPK((WikiPage)classedModel);
 	}
 
 	@Override
 	protected BaseModel<?> getBaseModel(long primaryKey) throws Exception {
-		return WikiNodeLocalServiceUtil.getNode(primaryKey);
+		return WikiPageTrashHandlerTestUtil.getBaseModel(primaryKey);
 	}
 
 	@Override
 	protected Class<?> getBaseModelClass() {
-		return WikiNode.class;
+		return WikiPageTrashHandlerTestUtil.getBaseModelClass();
 	}
 
 	@Override
 	protected String getBaseModelName(ClassedModel classedModel) {
-		WikiNode node = (WikiNode)classedModel;
-
-		return node.getName();
+		return WikiPageTrashHandlerTestUtil.getBaseModelName(classedModel);
 	}
 
 	@Override
 	protected int getNotInTrashBaseModelsCount(BaseModel<?> parentBaseModel)
 		throws Exception {
 
-		return WikiNodeLocalServiceUtil.getNodesCount(
-			(Long)parentBaseModel.getPrimaryKeyObj(),
-			WorkflowConstants.STATUS_APPROVED);
+		return WikiPageTrashHandlerTestUtil.getNotInTrashBaseModelsCount(
+			parentBaseModel);
+	}
+
+	@Override
+	protected BaseModel<?> getParentBaseModel(
+			Group group, ServiceContext serviceContext)
+		throws Exception {
+
+		return WikiPageTrashHandlerTestUtil.getParentBaseModel(
+			group, serviceContext);
+	}
+
+	@Override
+	protected Class<?> getParentBaseModelClass() {
+		return WikiNode.class;
 	}
 
 	@Override
 	protected String getSearchKeywords() {
-		return _NODE_NAME;
+		return WikiPageTrashHandlerTestUtil.getSearchKeywords();
+	}
+
+	@Override
+	protected long getTrashEntryClassPK(ClassedModel classedModel) {
+		return WikiPageTrashHandlerTestUtil.getTrashEntryClassPK(classedModel);
 	}
 
 	@Override
 	protected String getUniqueTitle(BaseModel<?> baseModel) {
-		WikiNode node = (WikiNode)baseModel;
-
-		return TrashUtil.getOriginalTitle(node.getName());
+		return WikiPageTrashHandlerTestUtil.getUniqueTitle(baseModel);
 	}
 
 	@Override
-	protected boolean isAssetableModel() {
-		return false;
-	}
-
-	@Override
-	protected boolean isIndexableBaseModel() {
+	protected boolean isBaseModelMoveableFromTrash() {
 		return false;
 	}
 
 	@Override
 	protected void moveBaseModelToTrash(long primaryKey) throws Exception {
-		WikiNodeLocalServiceUtil.moveNodeToTrash(
-			TestPropsValues.getUserId(), primaryKey);
+		WikiPageTrashHandlerTestUtil.moveBaseModelToTrash(primaryKey);
 	}
 
-	private static final String _NODE_NAME = RandomTestUtil.randomString(75);
+	@Override
+	protected void moveParentBaseModelToTrash(long primaryKey)
+		throws Exception {
+
+		WikiPageTrashHandlerTestUtil.moveParentBaseModelToTrash(primaryKey);
+	}
+
+	@Override
+	protected BaseModel<?> updateBaseModel(
+			long primaryKey, ServiceContext serviceContext)
+		throws Exception {
+
+		return WikiPageTrashHandlerTestUtil.updateBaseModel(
+			primaryKey, serviceContext);
+	}
 
 }
